refactor : 템플릿 메서드 패턴 적용 및 각 Controller의 미사용 추상함수는 Custom Exception을 Throw 하도록 수정, 미사용 메서드 삭제
