+++ src/main/java/controller/AbstractController.java
@@ -6,15 +6,15 @@ import webserver.Response;
 public abstract class AbstractController implements Controller {
     @Override
     public void service(Request request, Response response) throws Exception {
-
+        if (request.getMethod().equals("GET")) {
+            doGet(request, response);
+        }
+        if (request.getMethod().equals("POST")) {
+            doPost(request, response);
+        }
     }
 
-    public void doPost(Request request, Response response)  throws Exception {
+    public abstract void doPost(Request request, Response response) throws Exception;
 
-    }
-
-
-    public void doGet(Request request, Response response) throws Exception {
-
-    }
+    public abstract void doGet(Request request, Response response) throws Exception;
 }
