+++ src/main/java/controller/CreateUserController.java
@@ -1,20 +1,12 @@
 package controller;
 
 import db.DataBase;
+import exception.NotDefinedMethodException;
 import model.User;
 import webserver.Request;
 import webserver.Response;
 
 public class CreateUserController extends AbstractController {
-    @Override
-    public void service(Request request, Response response) throws Exception {
-        if (request.getMethod().equals("GET")) {
-            doGet(request, response);
-        }
-        if (request.getMethod().equals("POST")) {
-            doPost(request, response);
-        }
-    }
 
     @Override
     public void doPost(Request request, Response response) throws Exception {
@@ -22,4 +14,10 @@ public class CreateUserController extends AbstractController {
         DataBase.addUser(user);
         response.sendRedirect("/index.html");
     }
+
+    @Override
+    public void doGet(Request request, Response response) throws Exception {
+        throw new NotDefinedMethodException();
+    }
+
 }
