+++ src/main/java/controller/ListUserController.java
@@ -1,19 +1,10 @@
 package controller;
 
+import exception.NotDefinedMethodException;
 import webserver.Request;
 import webserver.Response;
 
 public class ListUserController extends AbstractController {
-    @Override
-    public void service(Request request, Response response) throws Exception {
-        if (request.getMethod().equals("GET")) {
-            doGet(request, response);
-        }
-        if (request.getMethod().equals("POST")) {
-            doPost(request, response);
-        }
-    }
-
     @Override
     public void doGet(Request request, Response response) throws Exception {
         if (request.getHeader("Cookie") != null &&
@@ -22,5 +13,10 @@ public class ListUserController extends AbstractController {
             return;
         }
         response.sendRedirect("/user/login.html");
+    }
+
+    @Override
+    public void doPost(Request request, Response response) throws Exception {
+        throw new NotDefinedMethodException();
     }
 }
