--- src/main/java/utils/PathUtils.java
@@ -1,8 +1,0 @@
-package utils;
-
-public class PathUtils {
-
-    public static String parseGetURI(String token) {
-        return null;
-    }
-}
